edit customer and fares almost completed, there are unresolved issues:
- editing the address of the customer (is inside the role, not the user)
doesn't work once it's created... it doesn't give any error and even in
the view it works but when you end the conversation the changes are
gone.

- the changes in the fares are not registered in the customer view, but
when you end the conversation and select the customer again you can see
them even in the customer view... it seem like an opposite problem to
the address one.